Swapped icons for recording for practice mode, also adjuted the position of completion msg appears on the screen
