+++ AppCode/app/components/ActivityVideoControl.js
@@ -1,20 +1,19 @@
 import React from "react";
-import { StyleSheet, TouchableOpacity, View } from "react-native"
+import { StyleSheet, TouchableOpacity, View } from "react-native";
 import Icon from "./Icon";
 import colors from "../config/colors";
 
 function ActivityVideoControl({ onBack, onReplay }) {
-
   const replay = "refresh-circle"; //'replay';
-  const goBack = 'arrow-left';
+  const goBack = "arrow-left";
 
   const handleBack = () => {
     onBack();
-  }
+  };
 
   const handleReplay = () => {
     onReplay();
-  }
+  };
 
   return (
     <View style={styles.toolbar}>
@@ -22,7 +21,7 @@ function ActivityVideoControl({ onBack, onReplay }) {
         <Icon
           name={goBack}
           iconColor="#4b62f2"
-          backgroudColor="transparent"
+          backgroundColor="transparent"
           size={80}
           style={styles.button}
         />
@@ -31,34 +30,32 @@ function ActivityVideoControl({ onBack, onReplay }) {
         <Icon
           name={replay}
           iconColor={colors.darkblue}
-          backgroudColor="transparent"
+          backgroudColor={colors.transparent}
           size={160}
           style={styles.button}
         />
       </TouchableOpacity>
     </View>
-  )
+  );
 }
 
 const styles = StyleSheet.create({
   toolbar: {
-    position: 'absolute',
-    flexDirection: 'row',
-    width: '100%',
-    height: '100%',
-    justifyContent: 'center',
+    flex: 1,
+    justifyContent: "center",
+    backgroundColor: colors.transparent,
   },
   buttonContainer: {
-    backgroundColor: "white",
-    alignItems: 'center',
-    justifyContent: 'center',
+    backgroundColor: colors.white,
+    alignItems: "center",
+    justifyContent: "center",
     borderRadius: 30,
     width: 60,
     height: 60,
   },
   button: {
-    alignSelf: 'center',
-  }
-})
+    alignSelf: "center",
+  },
+});
 
 export default ActivityVideoControl;
